requirements.txt
 flask
 flask-smorest
 python-dotenv
 flask-sqlalchemy